refactor - #9 코드리뷰 반영
