+++ src/main/java/sopt/org/HMH/domain/challenge/controller/ChallengeController.java
@@ -15,14 +15,14 @@ import sopt.org.HMH.global.common.response.ApiResponse;
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/challenge")
 public class ChallengeController {
+
     private final ChallengeService challengeService;
 
     @PostMapping
     public ResponseEntity<ApiResponse> orderAdd(@RequestBody ChallengeRequest request) {
-        // db에 임시 유저 (userId=1) 존재 (임시 사용)
-        Long response = challengeService.add(1L, request);
+        // TODO: - 토큰으로 유저 아이디 찾는 함수 연결
         return ResponseEntity
                 .status(ChallengeSuccess.SUCCESS_CREATE_CHALLENGE.getHttpStatus())
-                .body(ApiResponse.success(ChallengeSuccess.SUCCESS_CREATE_CHALLENGE, response));
+                .body(ApiResponse.success(ChallengeSuccess.SUCCESS_CREATE_CHALLENGE, challengeService.add(1L, request));
     }
 }
