+++ src/main/java/sopt/org/HMH/domain/user/User.java
@@ -1,7 +1,10 @@
 package sopt.org.HMH.domain.user;
 
 import jakarta.persistence.*;
+import sopt.org.HMH.domain.challenge.domain.Challenge;
 import sopt.org.HMH.global.common.domain.BaseTimeEntity;
+
+import java.util.List;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
 
@@ -13,4 +16,7 @@ public class User extends BaseTimeEntity {
     @GeneratedValue(strategy = IDENTITY)
     @Column(name = "user_id")
     private Integer Id;
+
+    @OneToMany(mappedBy = "user")
+    private List<Challenge> challenges;
 }
