+++ src/main/java/sopt/org/HMH/global/common/domain/BaseTimeEntity.java
@@ -14,9 +14,8 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 public abstract class BaseTimeEntity {
 
     @CreatedDate
-    protected LocalDateTime createdAt;
+    private LocalDateTime createdAt;
 
     @LastModifiedDate
-    protected LocalDateTime updatedAt;
-
+    private LocalDateTime updatedAt;
 }
